pyproject.toml
[project]
name = "flight-calendar-updater"
version = "0.1.0"
description = ""
authors = [
    {name = "momonala", email = "mnalavadi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aniso8601>=10.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "openai>=1.0.0",
    "joblib>=1.0.0",
    "google-api-python-client>=2.0.0,<3.0.0",
    "google-auth-httplib2>=0.0.0,<1.0.0",
    "google-auth-oauthlib>=0.0.0,<1.0.0",
    "pandas",
    "requests>=2.0.0,<3.0.0",
    "selenium>=4.6.0,<5.0.0",
    "beautifulsoup4>=4.0.0,<5.0.0",
    "airportsdata",
    "country-converter>=1.0.0,<2.0.0",
    "pycountry>=24.0.0,<25.0.0",
    "pytz",
    "schedule",
    "brotli>=1.2.0",
    "zstandard>=0.25.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black",
    "jupyter>=1.1.1",
    "notebook>=7.5.1",
    "pandas>=2.3.3",
    "ruff>=0.14.10",
    "isort>=7.0.0",
    "typer>=0.9.0",
    "asyncio>=4.0.0",
    "python-dotenv>=1.2.2",
]

[tool.config]
# Non-secret config; secrets (SPREADSHEET_ID, CALENDAR_ID, OPENAI_API_KEY) come from env vars
openai_model = "gpt-5.4"
scheduler_trigger_time = "10:00"
range_name = "raw!A:ZZ"
service_account_file = "google_application_credentials.json"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
main = "src.main:app"
config = "src.config:main"


[tool.black]
line-length = 110
target-version = ['py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-s -v --cov=src --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.uv]
exclude-newer = "1 week"
